+++ cypress/integration/admin/professions/publish.spec.ts
@@ -139,6 +139,9 @@ describe('Publishing professions', () => {
         'professions.form.label.qualifications.routesToObtain',
       );
       cy.checkAccessibility();
+      cy.get(
+        'input[name="otherCountriesRecognitionRoutes"][value="some"]',
+      ).check();
       cy.get('textarea[name="routesToObtain"]').type('Routes to obtain');
       cy.translate('app.continue').then((buttonText) => {
         cy.get('button').contains(buttonText).click();
@@ -474,8 +477,17 @@ describe('Publishing professions', () => {
         'http://example.com/more-info',
       );
 
-      cy.get('input[name="ukRecognition"]').type('Recognition in the UK');
+      cy.get('textarea[name="ukRecognition"]').type('Recognition in the UK');
       cy.get('input[name="ukRecognitionUrl"]').type('http://example.com/uk');
+      cy.get(
+        'input[name="otherCountriesRecognitionRoutes"][value="some"]',
+      ).check();
+      cy.get('textarea[name="otherCountriesRecognitionSummary"]').type(
+        'Recognition in other countries',
+      );
+      cy.get('input[name="otherCountriesRecognitionUrl"]').type(
+        'http://example.com/other',
+      );
 
       cy.translate('app.continue').then((buttonText) => {
         cy.get('button').contains(buttonText).click();