+++ src/professions/admin/qualifications.controller.spec.ts
@@ -17,6 +17,7 @@ import { translationOf } from '../../testutils/translation-of';
 import userFactory from '../../testutils/factories/user';
 import { createDefaultMockRequest } from '../../testutils/factories/create-default-mock-request';
 import { checkCanViewProfession } from '../../users/helpers/check-can-view-profession';
+import { OtherCountriesRecognitionRoutes } from '../../qualifications/qualification.entity';
 
 jest.mock('../../helpers/nations.helper');
 jest.mock('../../users/helpers/check-can-view-profession');
@@ -81,6 +82,10 @@ describe(QualificationsController, () => {
             captionText: translationOf('professions.form.captions.edit'),
             ukRecognition: profession.qualification.ukRecognition,
             ukRecognitionUrl: profession.qualification.ukRecognitionUrl,
+            otherCountriesRecognitionSummary:
+              profession.qualification.otherCountriesRecognitionSummary,
+            otherCountriesRecognitionUrl:
+              profession.qualification.ukRecognitionUrl,
             isUK: false,
           }),
         );
@@ -158,6 +163,11 @@ describe(QualificationsController, () => {
           moreInformationUrl: 'http://www.example.com/more-info',
           ukRecognition: 'ukRecognition',
           ukRecognitionUrl: 'http://example.com/uk',
+          otherCountriesRecognitionRoutes:
+            'all' as OtherCountriesRecognitionRoutes,
+          otherCountriesRecognitionSummary:
+            'other countries recognition summary',
+          otherCountriesRecognitionUrl: 'http://example.com/overseas',
         };
 
         professionsService.findWithVersions.mockResolvedValue(profession);
@@ -182,6 +192,11 @@ describe(QualificationsController, () => {
               url: 'http://www.example.com/more-info',
               ukRecognition: 'ukRecognition',
               ukRecognitionUrl: 'http://example.com/uk',
+              otherCountriesRecognitionRoutes:
+                'all' as OtherCountriesRecognitionRoutes,
+              otherCountriesRecognitionSummary:
+                'other countries recognition summary',
+              otherCountriesRecognitionUrl: 'http://example.com/overseas',
             }),
           }),
         );
@@ -207,6 +222,11 @@ describe(QualificationsController, () => {
           moreInformationUrl: 'www.example.com/more-info ',
           ukRecognition: 'ukRecognition',
           ukRecognitionUrl: 'example.com/uk',
+          otherCountriesRecognitionRoutes:
+            'all' as OtherCountriesRecognitionRoutes,
+          otherCountriesRecognitionSummary:
+            'other countries recognition summary',
+          otherCountriesRecognitionUrl: 'example.com/overseas',
         };
 
         professionsService.findWithVersions.mockResolvedValue(profession);
@@ -231,6 +251,11 @@ describe(QualificationsController, () => {
               url: 'http://www.example.com/more-info',
               ukRecognition: 'ukRecognition',
               ukRecognitionUrl: 'http://example.com/uk',
+              otherCountriesRecognitionRoutes:
+                'all' as OtherCountriesRecognitionRoutes,
+              otherCountriesRecognitionSummary:
+                'other countries recognition summary',
+              otherCountriesRecognitionUrl: 'http://example.com/overseas',
             }),
           }),
         );
@@ -256,6 +281,9 @@ describe(QualificationsController, () => {
           moreInformationUrl: 'not a url',
           ukRecognition: '',
           ukRecognitionUrl: 'not a url',
+          otherCountriesRecognitionRoutes: undefined,
+          otherCountriesRecognitionSummary: '',
+          otherCountriesRecognitionUrl: 'not a url',
         };
 
         professionsService.findWithVersions.mockResolvedValue(profession);
@@ -288,6 +316,12 @@ describe(QualificationsController, () => {
               ukRecognitionUrl: {
                 text: 'professions.form.errors.qualification.ukRecognitionUrl.invalid',
               },
+              otherCountriesRecognitionRoutes: {
+                text: 'professions.form.errors.qualification.otherCountriesRecognitionRoutes.empty',
+              },
+              otherCountriesRecognitionUrl: {
+                text: 'professions.form.errors.qualification.otherCountriesRecognitionUrl.invalid',
+              },
             },
           }),
         );
@@ -308,6 +342,10 @@ describe(QualificationsController, () => {
         moreInformationUrl: 'http://www.example.com/more-info',
         ukRecognition: 'ukRecognition',
         ukRecognitionUrl: 'http://example.com/uk',
+        otherCountriesRecognitionRoutes:
+          'all' as OtherCountriesRecognitionRoutes,
+        otherCountriesRecognitionSummary: 'other countries recognition summary',
+        otherCountriesRecognitionUrl: 'http://example.com/overseas',
       };
 
       await controller.update(