+++ src/qualifications/presenters/qualification.presenter.spec.ts
@@ -6,6 +6,7 @@ import { createMockI18nService } from '../../testutils/create-mock-i18n-service'
 import { translationOf } from '../../testutils/translation-of';
 import { formatLink } from '../../helpers/format-link.helper';
 import { linkOf } from '../../testutils/link-of';
+import { OtherCountriesRecognitionRoutes } from '../qualification.entity';
 
 jest.mock('../../helpers/format-multiline-string.helper');
 jest.mock('../../helpers/format-link.helper');
@@ -47,6 +48,80 @@ describe(QualificationPresenter, () => {
         expect(presenter.moreInformationUrl).toEqual(
           linkOf('http://example.com'),
         );
+      });
+    });
+
+    describe('adminSelectedOtherCountriesRecognitionRoutes', () => {
+      describe('when other routes are set', () => {
+        it('returnsthe route', () => {
+          const qualification = qualificationFactory.build({
+            otherCountriesRecognitionRoutes:
+              OtherCountriesRecognitionRoutes.All,
+          });
+
+          const presenter = new QualificationPresenter(
+            qualification,
+            createMockI18nService(),
+          );
+
+          expect(
+            presenter.adminSelectedOtherCountriesRecognitionRoutes,
+          ).toEqual(OtherCountriesRecognitionRoutes.All);
+        });
+      });
+
+      describe('when other countries routes are not set', () => {
+        it('returns null', () => {
+          const qualification = qualificationFactory.build({
+            otherCountriesRecognitionRoutes: null,
+          });
+
+          const presenter = new QualificationPresenter(
+            qualification,
+            createMockI18nService(),
+          );
+
+          expect(
+            presenter.adminSelectedOtherCountriesRecognitionRoutes,
+          ).toEqual(null);
+        });
+      });
+    });
+
+    describe('publicOtherCountriesRecognitionRoutes', () => {
+      describe('when other routes are set', () => {
+        it('returns a localisation id with the route', () => {
+          const qualification = qualificationFactory.build({
+            otherCountriesRecognitionRoutes:
+              OtherCountriesRecognitionRoutes.All,
+          });
+
+          const presenter = new QualificationPresenter(
+            qualification,
+            createMockI18nService(),
+          );
+
+          expect(presenter.publicOtherCountriesRecognitionRoutes).toEqual(
+            OtherCountriesRecognitionRoutes.All,
+          );
+        });
+      });
+
+      describe('when other countries routes are not set', () => {
+        it('returns null', () => {
+          const qualification = qualificationFactory.build({
+            otherCountriesRecognitionRoutes: null,
+          });
+
+          const presenter = new QualificationPresenter(
+            qualification,
+            createMockI18nService(),
+          );
+
+          expect(
+            presenter.adminSelectedOtherCountriesRecognitionRoutes,
+          ).toEqual(null);
+        });
       });
     });
 
@@ -110,6 +185,38 @@ describe(QualificationPresenter, () => {
                     html: presenter.moreInformationUrl,
                   },
                 },
+                {
+                  key: {
+                    text: translationOf(
+                      'professions.show.qualification.otherCountriesRecognition.routes.label',
+                    ),
+                  },
+                  value: {
+                    text: translationOf(
+                      `professions.show.qualification.otherCountriesRecognition.routes.${presenter.publicOtherCountriesRecognitionRoutes}`,
+                    ),
+                  },
+                },
+                {
+                  key: {
+                    text: translationOf(
+                      'professions.show.qualification.otherCountriesRecognition.summary',
+                    ),
+                  },
+                  value: {
+                    html: presenter.otherCountriesRecognitionSummary,
+                  },
+                },
+                {
+                  key: {
+                    text: translationOf(
+                      'professions.show.qualification.otherCountriesRecognition.url',
+                    ),
+                  },
+                  value: {
+                    html: presenter.otherCountriesRecognitionUrl,
+                  },
+                },
               ],
             });
           });
@@ -163,7 +270,7 @@ describe(QualificationPresenter, () => {
                     ),
                   },
                   value: {
-                    text: presenter.ukRecognition,
+                    html: presenter.ukRecognition,
                   },
                 },
                 {
@@ -174,6 +281,38 @@ describe(QualificationPresenter, () => {
                   },
                   value: {
                     html: presenter.ukRecognitionUrl,
+                  },
+                },
+                {
+                  key: {
+                    text: translationOf(
+                      'professions.show.qualification.otherCountriesRecognition.routes.label',
+                    ),
+                  },
+                  value: {
+                    text: translationOf(
+                      `professions.show.qualification.otherCountriesRecognition.routes.${presenter.publicOtherCountriesRecognitionRoutes}`,
+                    ),
+                  },
+                },
+                {
+                  key: {
+                    text: translationOf(
+                      'professions.show.qualification.otherCountriesRecognition.summary',
+                    ),
+                  },
+                  value: {
+                    html: presenter.otherCountriesRecognitionSummary,
+                  },
+                },
+                {
+                  key: {
+                    text: translationOf(
+                      'professions.show.qualification.otherCountriesRecognition.url',
+                    ),
+                  },
+                  value: {
+                    html: presenter.otherCountriesRecognitionUrl,
                   },
                 },
               ],
@@ -224,6 +363,38 @@ describe(QualificationPresenter, () => {
                     html: presenter.moreInformationUrl,
                   },
                 },
+                {
+                  key: {
+                    text: translationOf(
+                      'professions.show.qualification.otherCountriesRecognition.routes.label',
+                    ),
+                  },
+                  value: {
+                    text: translationOf(
+                      `professions.show.qualification.otherCountriesRecognition.routes.${presenter.publicOtherCountriesRecognitionRoutes}`,
+                    ),
+                  },
+                },
+                {
+                  key: {
+                    text: translationOf(
+                      'professions.show.qualification.otherCountriesRecognition.summary',
+                    ),
+                  },
+                  value: {
+                    html: presenter.otherCountriesRecognitionSummary,
+                  },
+                },
+                {
+                  key: {
+                    text: translationOf(
+                      'professions.show.qualification.otherCountriesRecognition.url',
+                    ),
+                  },
+                  value: {
+                    html: presenter.otherCountriesRecognitionUrl,
+                  },
+                },
               ],
             });
           });
@@ -259,6 +430,18 @@ describe(QualificationPresenter, () => {
                     html: presenter.routesToObtain,
                   },
                 },
+                {
+                  key: {
+                    text: translationOf(
+                      'professions.show.qualification.otherCountriesRecognition.routes.label',
+                    ),
+                  },
+                  value: {
+                    text: translationOf(
+                      `professions.show.qualification.otherCountriesRecognition.routes.${presenter.publicOtherCountriesRecognitionRoutes}`,
+                    ),
+                  },
+                },
               ],
             });
           });
@@ -279,8 +462,12 @@ describe(QualificationPresenter, () => {
           routesToObtain: multilineOf(undefined),
           moreInformationUrl: linkOf(undefined),
           qualification: undefined,
-          ukRecognition: undefined,
+          ukRecognition: multilineOf(undefined),
           ukRecognitionUrl: linkOf(undefined),
+          adminSelectedOtherCountriesRecognitionRoutes: undefined,
+          publicOtherCountriesRecognitionRoutes: undefined,
+          otherCountriesRecognitionSummary: multilineOf(undefined),
+          otherCountriesRecognitionUrl: linkOf(undefined),
         }),
       );
     });