+++ CHANGELOG.md
@@ -29,6 +29,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - New landing page for central users
+- Add back other countries routes to recognition with new radio buttons
 
 ### Changed